moved lexer, parser and token packages to service. removed counter and evaluator packages from service
