+++ src/main/java/com/example/myTestApp/controller/TestAppRestController.java
@@ -4,10 +4,8 @@ package com.example.myTestApp.controller;
 import com.example.myTestApp.exception.ExpressionNotFoundException;
 import com.example.myTestApp.model.Expression;
 import com.example.myTestApp.repository.ExpressionRepository;
-import com.example.myTestApp.service.evaluator.ArithmeticExpressionEvaluator;
-import com.example.myTestApp.service.counter.NumOfDoublesCounter;
-import com.example.myTestApp.service.updatedevaluator.lexer.Lexer;
-import com.example.myTestApp.service.updatedevaluator.parser.Parser;
+import com.example.myTestApp.service.lexer.Lexer;
+import com.example.myTestApp.service.parser.Parser;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
