--- src/main/java/com/example/myTestApp/service/counter/NumOfDoublesCounter.java
@@ -1,5 +1,0 @@
-package com.example.myTestApp.service.counter;
-
-public interface NumOfDoublesCounter {
-    int count (String str);
-}
