--- src/main/java/com/example/myTestApp/service/counter/NumOfDoublesCounterImpl.java
@@ -1,24 +1,0 @@
-package com.example.myTestApp.service.counter;
-
-import com.example.myTestApp.service.counter.NumOfDoublesCounter;
-import org.springframework.stereotype.Service;
-
-@Service
-public class NumOfDoublesCounterImpl implements NumOfDoublesCounter {
-    @Override
-    public int count(String str) {
-        var arr = str.toCharArray();
-        var num = 0;
-        var isNum = false;
-        for (var i = 0; i < arr.length; i++) {
-            if (Character.isDigit(arr[i])) {
-                isNum = true;
-            }
-            if (isNum && ((!Character.isDigit(arr[i]) && arr[i] != '.') || i == arr.length - 1)) {
-                num++;
-                isNum = false;
-            }
-        }
-        return num;
-    }
-}
