--- src/main/java/com/example/myTestApp/service/evaluator/ArithmeticExpressionEvaluator.java
@@ -1,8 +1,0 @@
-package com.example.myTestApp.service.evaluator;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public interface ArithmeticExpressionEvaluator {
-    double getEvaluation(String arithmeticExpression);
-}
