+++ src/main/java/com/example/myTestApp/service/parser/ParserImpl.java
@@ -1,7 +1,7 @@
-package com.example.myTestApp.service.updatedevaluator.parser;
+package com.example.myTestApp.service.parser;
 
 import com.example.myTestApp.exception.*;
-import com.example.myTestApp.service.updatedevaluator.token.*;
+import com.example.myTestApp.service.token.*;
 import org.springframework.stereotype.Service;
 
 import java.util.LinkedList;
