+++ src/main/java/com/example/myTestApp/service/token/Token.java
@@ -1,0 +1,4 @@
+package com.example.myTestApp.service.token;
+
+public interface Token {
+}
