--- src/main/java/com/example/myTestApp/service/updatedevaluator/lexer/Lexer.java
@@ -1,9 +1,0 @@
-package com.example.myTestApp.service.updatedevaluator.lexer;
-
-import com.example.myTestApp.service.updatedevaluator.token.Token;
-
-import java.util.List;
-
-public interface Lexer {
-    List<Token> tokenize (String receivedString);
-}
