--- src/main/java/com/example/myTestApp/service/updatedevaluator/token/Token.java
@@ -1,4 +1,0 @@
-package com.example.myTestApp.service.updatedevaluator.token;
-
-public interface Token {
-}
